+++ python/cuml/preprocessing/text/stem/porter_stemmer_utils/len_flags_utils.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 #
 
+
 def len_gt_n(word_strs, n):
     return word_strs.str.len() > n