+++ python/cuml/test/stemmer_tests/test_suffix_utils.py
@@ -15,21 +15,30 @@
 #
 
 import cudf
-from cuml.preprocessing.text.stem.porter_stemmer_utils.suffix_utils import get_stem_series, replace_suffix
+from cuml.preprocessing.text.stem.porter_stemmer_utils.suffix_utils import (
+    get_stem_series,
+    replace_suffix,
+)
 
 
 def test_get_stem_series():
-    word_strs = cudf.Series(["ihop", "packit", "mishit", "crow", "girl", "boy"])
+    word_strs = cudf.Series(
+        ["ihop", "packit", "mishit", "crow", "girl", "boy"]
+    )
     can_replace_mask = cudf.Series([True, True, True, False, False, False])
 
     expect = ["ih", "pack", "mish", "crow", "girl", "boy"]
-    got = get_stem_series(word_strs, suffix_len=2, can_replace_mask=can_replace_mask)
+    got = get_stem_series(
+        word_strs, suffix_len=2, can_replace_mask=can_replace_mask
+    )
     assert sorted(list(got.to_pandas().values)) == sorted(expect)
 
 
 def test_replace_suffix():
     # test 'ing' -> 's'
-    word_strs = cudf.Series(["shopping", "parking", "drinking", "sing", "bing"])
+    word_strs = cudf.Series(
+        ["shopping", "parking", "drinking", "sing", "bing"]
+    )
     can_replace_mask = cudf.Series([True, True, True, False, False])
     got = replace_suffix(word_strs, "ing", "s", can_replace_mask)
     expect = ["shopps", "parks", "drinks", "sing", "bing"]